feat: remove Rick and Morty api and screens
