--- composeApp/src/commonMain/kotlin/data/characters/CharacterWrapper.kt
@@ -1,11 +1,0 @@
-package data.characters
-
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CharacterWrapper(
-    @SerialName("results")
-    val results: List<Character>
-)
