--- composeApp/src/commonMain/kotlin/data/ktor/KtorClient.kt
@@ -1,6 +1,0 @@
-package data.ktor
-
-import io.ktor.client.HttpClient
-import io.ktor.client.HttpClientConfig
-
-expect fun ktorClient(block: HttpClientConfig<*>.() -> Unit = {}) : HttpClient
