--- composeApp/src/iosMain/kotlin/MainViewController.kt
@@ -1,6 +1,0 @@
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() {
-    TODO()
-//    ComposeUIViewController { App() }
-}
