--- composeApp/src/iosMain/kotlin/data/ktor/KtorClient.ios.kt
@@ -1,8 +1,0 @@
-package data.ktor
-
-import io.ktor.client.HttpClient
-import io.ktor.client.HttpClientConfig
-
-actual fun ktorClient(block: HttpClientConfig<*>.() -> Unit): HttpClient {
-    TODO("Not yet implemented")
-}
