--- composeApp/src/wasmJsMain/kotlin/data/ktor/KtorClient.wasmJs.kt
@@ -1,8 +1,0 @@
-package data.ktor
-
-import io.ktor.client.HttpClient
-import io.ktor.client.HttpClientConfig
-
-actual fun ktorClient(block: HttpClientConfig<*>.() -> Unit): HttpClient {
-    return HttpClient(block)
-}
